Fix 232 non-matching functions in batch_c/d files
Three categories of fixes applied:
- Return type: u64→u32 for 99 functions (32-bit mov w0,wzr)
- Register variables: register x8/x10 asm() for 221 vtable dispatch
  functions where Ghidra missed live-in register usage
- Post-processing: fix_prologue.py (NX frame pointer scheduling),
  fix_epilogue.py (NX epilogue ordering via whitelist)

batch_c/d: 10→242 verified (+232)
Overall: 5884→6141 verified (+257)

+++ src/app/fun_batch_c_010.cpp
@@ -126,7 +126,7 @@ u64 FUN_71021c4000(app::BattleObjectModuleAccessor *param_1, s64 param_2)
 }
 
 // 0x71021c7c1c
-u64 FUN_71021c7c1c(app::BattleObjectModuleAccessor *param_1, s64 param_2, bool param_3, bool param_4)
+u32 FUN_71021c7c1c(app::BattleObjectModuleAccessor *param_1, s64 param_2, bool param_3, bool param_4)
 {
     u64 in_x10;
     *(u64 *)(param_2 + 0x10) = in_x10;
@@ -252,7 +252,7 @@ u64 FUN_71021e1290(app::Fighter *param_1, s64 param_2, bool param_3)
 }
 
 // 0x71021e1690
-u64 FUN_71021e1690(s64 param_1, s64 param_2)
+u32 FUN_71021e1690(s64 param_1, s64 param_2)
 {
     s64 lVar1;
     u64 in_x10;
@@ -282,14 +282,14 @@ u64 FUN_71021f6514(s64 *param_1)
 }
 
 // 0x71021f72f0
-u64 FUN_71021f72f0(void)
+u32 FUN_71021f72f0(void)
 {
     FUN_71015c0410();
     return 0;
 }
 
 // 0x71021fabc4
-u64 FUN_71021fabc4(s64 param_1)
+u32 FUN_71021fabc4(s64 param_1)
 {
     s64 *plVar1 = *(s64 **)(param_1 + 0x208);
     (**(void (**)())(*plVar1 + 0xc8))();
@@ -297,14 +297,14 @@ u64 FUN_71021fabc4(s64 param_1)
 }
 
 // 0x71021fbaa0
-u64 FUN_71021fbaa0(s64 param_1)
+u32 FUN_71021fbaa0(s64 param_1)
 {
     FUN_71015b08c0(*(u64 *)(*(s64 *)(*(s64 *)(*(s64 *)(param_1 - 8) + 0x1a0) + 0x190) + 0x220));
     return 0;
 }
 
 // 0x71021fc278
-u64 FUN_71021fc278(s64 param_1, s64 *param_2)
+u32 FUN_71021fc278(s64 param_1, s64 *param_2)
 {
     *(u16 *)(param_1 + 0x150) = 1;
     *(u8 *)(param_1 + 0x152) = 0;
@@ -313,28 +313,28 @@ u64 FUN_71021fc278(s64 param_1, s64 *param_2)
 }
 
 // 0x71021ff030
-u64 FUN_71021ff030(lua_State *param_1)
+u32 FUN_71021ff030(lua_State *param_1)
 {
     app::item::apply_powerup_throw(param_1);
     return 0;
 }
 
 // 0x71021ff1c0
-u64 FUN_71021ff1c0(lua_State *param_1)
+u32 FUN_71021ff1c0(lua_State *param_1)
 {
     app::item::apply_powerup_etc(param_1);
     return 0;
 }
 
 // 0x71021ff230
-u64 FUN_71021ff230(lua_State *param_1)
+u32 FUN_71021ff230(lua_State *param_1)
 {
     app::item::apply_powerup_attr(param_1);
     return 0;
 }
 
 // 0x71022004e0
-u64 FUN_71022004e0(s64 param_1)
+u32 FUN_71022004e0(s64 param_1)
 {
     s64 *plVar1;
     plVar1 = *(s64 **)(*(s64 *)(*(s64 *)(*(s64 *)(*(s64 *)(param_1 - 8) + 0x1a0) + 0x190) + 0x220) + 0x100);
@@ -343,7 +343,7 @@ u64 FUN_71022004e0(s64 param_1)
 }
 
 // 0x7102203b00
-u64 FUN_7102203b00(s64 param_1)
+u32 FUN_7102203b00(s64 param_1)
 {
     FUN_71015b3a40(*(u64 *)(*(s64 *)(*(s64 *)(*(s64 *)(param_1 - 8) + 0x1a0) + 0x190) + 0x220));
     return 0;
@@ -358,14 +358,14 @@ u64 FUN_7102208ccc(s64 param_1, s64 param_2)
 }
 
 // 0x7102208d00
-u64 FUN_7102208d00(lua_State *param_1)
+u32 FUN_7102208d00(lua_State *param_1)
 {
     app::boss_private::send_event_on_boss_dead(param_1);
     return 0;
 }
 
 // 0x7102208df0
-u64 FUN_7102208df0(s64 param_1)
+u32 FUN_7102208df0(s64 param_1)
 {
     s64 *plVar1;
     plVar1 = *(s64 **)(*(s64 *)(*(s64 *)(param_1 - 8) + 0x1a0) + 0x78);
@@ -374,7 +374,7 @@ u64 FUN_7102208df0(s64 param_1)
 }
 
 // 0x710220bf50
-u64 FUN_710220bf50(lua_State *param_1)
+u32 FUN_710220bf50(lua_State *param_1)
 {
     app::item_collision_manager::disable_ground_collision(param_1);
     return 0;
