+++ src/app/fun_batch_c_012.cpp
@@ -4,7 +4,7 @@
 #include "types.h"
 
 // 0x7102208730
-u64 FUN_7102208730(s64 param_1)
+u32 FUN_7102208730(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -15,7 +15,7 @@ u64 FUN_7102208730(s64 param_1)
 }
 
 // 0x7102208770
-u64 FUN_7102208770(s64 param_1)
+u32 FUN_7102208770(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -26,7 +26,7 @@ u64 FUN_7102208770(s64 param_1)
 }
 
 // 0x7102208810
-u64 FUN_7102208810(s64 param_1)
+u32 FUN_7102208810(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -37,7 +37,7 @@ u64 FUN_7102208810(s64 param_1)
 }
 
 // 0x7102208850
-u64 FUN_7102208850(s64 param_1)
+u32 FUN_7102208850(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -48,7 +48,7 @@ u64 FUN_7102208850(s64 param_1)
 }
 
 // 0x710221c620
-u64 FUN_710221c620(s64 param_1)
+u32 FUN_710221c620(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -59,7 +59,7 @@ u64 FUN_710221c620(s64 param_1)
 }
 
 // 0x710221c700
-u64 FUN_710221c700(s64 param_1)
+u32 FUN_710221c700(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -70,7 +70,7 @@ u64 FUN_710221c700(s64 param_1)
 }
 
 // 0x710221e780
-u64 FUN_710221e780(s64 param_1)
+u32 FUN_710221e780(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -81,7 +81,7 @@ u64 FUN_710221e780(s64 param_1)
 }
 
 // 0x710221e860
-u64 FUN_710221e860(s64 param_1)
+u32 FUN_710221e860(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -92,7 +92,7 @@ u64 FUN_710221e860(s64 param_1)
 }
 
 // 0x71022208e0
-u64 FUN_71022208e0(s64 param_1)
+u32 FUN_71022208e0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -103,7 +103,7 @@ u64 FUN_71022208e0(s64 param_1)
 }
 
 // 0x71022209c0
-u64 FUN_71022209c0(s64 param_1)
+u32 FUN_71022209c0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -114,7 +114,7 @@ u64 FUN_71022209c0(s64 param_1)
 }
 
 // 0x7102220a00
-u64 FUN_7102220a00(s64 param_1)
+u32 FUN_7102220a00(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -125,7 +125,7 @@ u64 FUN_7102220a00(s64 param_1)
 }
 
 // 0x71022220e0
-u64 FUN_71022220e0(s64 param_1)
+u32 FUN_71022220e0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -136,7 +136,7 @@ u64 FUN_71022220e0(s64 param_1)
 }
 
 // 0x71022221c0
-u64 FUN_71022221c0(s64 param_1)
+u32 FUN_71022221c0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -147,7 +147,7 @@ u64 FUN_71022221c0(s64 param_1)
 }
 
 // 0x7102222200
-u64 FUN_7102222200(s64 param_1)
+u32 FUN_7102222200(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -158,7 +158,7 @@ u64 FUN_7102222200(s64 param_1)
 }
 
 // 0x7102222240
-u64 FUN_7102222240(s64 param_1)
+u32 FUN_7102222240(s64 param_1)
 {
     s64 *plVar1;
     plVar1 = *(s64 **)(*(s64 *)(*(s64 *)(param_1 - 8) + 0x1a0) + 0x68);
@@ -168,7 +168,7 @@ u64 FUN_7102222240(s64 param_1)
 }
 
 // 0x7102223ee0
-u64 FUN_7102223ee0(s64 param_1)
+u32 FUN_7102223ee0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -179,7 +179,7 @@ u64 FUN_7102223ee0(s64 param_1)
 }
 
 // 0x7102223fc0
-u64 FUN_7102223fc0(s64 param_1)
+u32 FUN_7102223fc0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -190,7 +190,7 @@ u64 FUN_7102223fc0(s64 param_1)
 }
 
 // 0x7102224000
-u64 FUN_7102224000(s64 param_1)
+u32 FUN_7102224000(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -201,7 +201,7 @@ u64 FUN_7102224000(s64 param_1)
 }
 
 // 0x7102225ae0
-u64 FUN_7102225ae0(s64 param_1)
+u32 FUN_7102225ae0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -212,7 +212,7 @@ u64 FUN_7102225ae0(s64 param_1)
 }
 
 // 0x7102225bc0
-u64 FUN_7102225bc0(s64 param_1)
+u32 FUN_7102225bc0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -223,7 +223,7 @@ u64 FUN_7102225bc0(s64 param_1)
 }
 
 // 0x7102227c00
-u64 FUN_7102227c00(s64 param_1)
+u32 FUN_7102227c00(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -234,7 +234,7 @@ u64 FUN_7102227c00(s64 param_1)
 }
 
 // 0x7102227ce0
-u64 FUN_7102227ce0(s64 param_1)
+u32 FUN_7102227ce0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -245,7 +245,7 @@ u64 FUN_7102227ce0(s64 param_1)
 }
 
 // 0x7102227d20
-u64 FUN_7102227d20(s64 param_1)
+u32 FUN_7102227d20(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -256,7 +256,7 @@ u64 FUN_7102227d20(s64 param_1)
 }
 
 // 0x7102229da0
-u64 FUN_7102229da0(s64 param_1)
+u32 FUN_7102229da0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -267,7 +267,7 @@ u64 FUN_7102229da0(s64 param_1)
 }
 
 // 0x7102229e80
-u64 FUN_7102229e80(s64 param_1)
+u32 FUN_7102229e80(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -278,7 +278,7 @@ u64 FUN_7102229e80(s64 param_1)
 }
 
 // 0x7102229ec0
-u64 FUN_7102229ec0(s64 param_1)
+u32 FUN_7102229ec0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -289,7 +289,7 @@ u64 FUN_7102229ec0(s64 param_1)
 }
 
 // 0x710222bde0
-u64 FUN_710222bde0(s64 param_1)
+u32 FUN_710222bde0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -300,7 +300,7 @@ u64 FUN_710222bde0(s64 param_1)
 }
 
 // 0x710222bec0
-u64 FUN_710222bec0(s64 param_1)
+u32 FUN_710222bec0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -311,7 +311,7 @@ u64 FUN_710222bec0(s64 param_1)
 }
 
 // 0x710222bf00
-u64 FUN_710222bf00(s64 param_1)
+u32 FUN_710222bf00(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -322,7 +322,7 @@ u64 FUN_710222bf00(s64 param_1)
 }
 
 // 0x710222cd00
-u64 FUN_710222cd00(s64 param_1)
+u32 FUN_710222cd00(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -333,7 +333,7 @@ u64 FUN_710222cd00(s64 param_1)
 }
 
 // 0x710222cde0
-u64 FUN_710222cde0(s64 param_1)
+u32 FUN_710222cde0(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -344,7 +344,7 @@ u64 FUN_710222cde0(s64 param_1)
 }
 
 // 0x710222ce20
-u64 FUN_710222ce20(s64 param_1)
+u32 FUN_710222ce20(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -355,7 +355,7 @@ u64 FUN_710222ce20(s64 param_1)
 }
 
 // 0x710222dc20
-u64 FUN_710222dc20(s64 param_1)
+u32 FUN_710222dc20(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -366,7 +366,7 @@ u64 FUN_710222dc20(s64 param_1)
 }
 
 // 0x710222dd00
-u64 FUN_710222dd00(s64 param_1)
+u32 FUN_710222dd00(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -377,7 +377,7 @@ u64 FUN_710222dd00(s64 param_1)
 }
 
 // 0x710222dd40
-u64 FUN_710222dd40(s64 param_1)
+u32 FUN_710222dd40(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -388,7 +388,7 @@ u64 FUN_710222dd40(s64 param_1)
 }
 
 // 0x710222fa90
-u64 FUN_710222fa90(s64 param_1)
+u32 FUN_710222fa90(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -399,7 +399,7 @@ u64 FUN_710222fa90(s64 param_1)
 }
 
 // 0x710222fb70
-u64 FUN_710222fb70(s64 param_1)
+u32 FUN_710222fb70(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -410,7 +410,7 @@ u64 FUN_710222fb70(s64 param_1)
 }
 
 // 0x7102230c40
-u64 FUN_7102230c40(s64 param_1)
+u32 FUN_7102230c40(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -421,7 +421,7 @@ u64 FUN_7102230c40(s64 param_1)
 }
 
 // 0x7102230d20
-u64 FUN_7102230d20(s64 param_1)
+u32 FUN_7102230d20(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
@@ -432,7 +432,7 @@ u64 FUN_7102230d20(s64 param_1)
 }
 
 // 0x7102230d60
-u64 FUN_7102230d60(s64 param_1)
+u32 FUN_7102230d60(s64 param_1)
 {
     s64 *plVar1;
     s64 lVar2;
