+++ src/app/fun_batch_d_005.cpp
@@ -104,37 +104,37 @@ u64 *FUN_7103791100(void)
 }
 
 // 0x7103716880 — nn::nfp::SetApplicationArea wrapper (256 bytes)
-u32 FUN_7103716880(s64 param_1, void *param_2, u64 param_3)
+u64 FUN_7103716880(s64 param_1, void *param_2, u64 param_3)
 {
     return nn::nfp::SetApplicationArea((nn::nfp::DeviceHandle*)(param_1 + 0x84), param_2, param_3);
 }
 
 // 0x7103716b80 — nn::nfp::GetTagInfo wrapper (256 bytes)
-u32 FUN_7103716b80(s64 param_1, nn::nfp::TagInfo *param_2)
+u64 FUN_7103716b80(s64 param_1, nn::nfp::TagInfo *param_2)
 {
     return nn::nfp::GetTagInfo(param_2, (nn::nfp::DeviceHandle*)(param_1 + 0x84));
 }
 
 // 0x7103716cd0 — nn::nfp::GetRegisterInfo wrapper (256 bytes)
-u32 FUN_7103716cd0(s64 param_1, nn::nfp::RegisterInfo *param_2)
+u64 FUN_7103716cd0(s64 param_1, nn::nfp::RegisterInfo *param_2)
 {
     return nn::nfp::GetRegisterInfo(param_2, (nn::nfp::DeviceHandle*)(param_1 + 0x84));
 }
 
 // 0x7103716e20 — nn::nfp::GetCommonInfo wrapper (256 bytes)
-u32 FUN_7103716e20(s64 param_1, nn::nfp::CommonInfo *param_2)
+u64 FUN_7103716e20(s64 param_1, nn::nfp::CommonInfo *param_2)
 {
     return nn::nfp::GetCommonInfo(param_2, (nn::nfp::DeviceHandle*)(param_1 + 0x84));
 }
 
 // 0x7103716f70 — nn::nfp::GetModelInfo wrapper (256 bytes)
-u32 FUN_7103716f70(s64 param_1, nn::nfp::ModelInfo *param_2)
+u64 FUN_7103716f70(s64 param_1, nn::nfp::ModelInfo *param_2)
 {
     return nn::nfp::GetModelInfo(param_2, (nn::nfp::DeviceHandle*)(param_1 + 0x84));
 }
 
 // 0x71037169d0 — nn::nfp::RecreateApplicationArea wrapper (256 bytes)
-u32 FUN_71037169d0(s64 param_1, nn::nfp::ApplicationAreaCreateInfo *param_2)
+u64 FUN_71037169d0(s64 param_1, nn::nfp::ApplicationAreaCreateInfo *param_2)
 {
     return nn::nfp::RecreateApplicationArea((nn::nfp::DeviceHandle*)(param_1 + 0x84), param_2);
 }
